# internal/docs/swagger.yaml
definitions:
  subscription.CUDResponse:
    properties:
      status:
        type: string
      subscription_id:
        type: string
    type: object
  subscription.CreateRequest:
    properties:
      end_date:
        type: string
      price:
        type: integer
      service_name:
        type: string
      start_date:
        type: string
      user_id:
        type: string
    type: object
  subscription.ListResponse:
    properties:
      subscriptions:
        items:
          $ref: '#/definitions/subscription.SubscriptionDTO'
        type: array
    type: object
  subscription.SubscriptionDTO:
    properties:
      end_date:
        type: string
      price:
        type: integer
      service_name:
        type: string
      start_date:
        type: string
      user_id:
        type: string
    type: object
  subscription.TotalCostResponse:
    properties:
      from:
        type: string
      service_name:
        type: string
      to:
        type: string
      total_cost:
        type: integer
      user_id:
        type: string
    type: object
  subscription.UpdateRequest:
    properties:
      end_date:
        type: string
      id:
        type: string
      price:
        type: integer
      service_name:
        type: string
      start_date:
        type: string
      user_id:
        type: string
    type: object
info:
  contact: {}
paths:
  /v1/healthz:
    get:
      description: Проверка, жив ли сервис (не зависит от БД)
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
      summary: Liveness probe
      tags:
      - health
  /v1/readyz:
    get:
      description: Проверка готовности сервиса (включая пинг базы данных)
      produces:
      - application/json
      responses:
        "200":
          description: ready
          schema:
            type: string
        "503":
          description: Service Unavailable
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Readiness probe
      tags:
      - health
  /v1/subscriptions:
    get:
      description: Получить список всех подписок
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/subscription.ListResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
        "504":
          description: Gateway Timeout
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List subscriptions
      tags:
      - subscriptions
    post:
      consumes:
      - application/json
      description: Создать новую подписку
      parameters:
      - description: Subscription payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/subscription.CreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/subscription.CUDResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
        "504":
          description: Gateway Timeout
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create subscription
      tags:
      - subscriptions
  /v1/subscriptions/{id}:
    delete:
      description: Удалить подписку по её идентификатору
      parameters:
      - description: Subscription ID (GUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/subscription.CUDResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
        "504":
          description: Gateway Timeout
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete subscription
      tags:
      - subscriptions
    get:
      description: Получить подписку по её идентификатору
      parameters:
      - description: Subscription ID (GUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/subscription.SubscriptionDTO'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
        "504":
          description: Gateway Timeout
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get subscription by ID
      tags:
      - subscriptions
    put:
      consumes:
      - application/json
      description: Обновить данные существующей подписки
      parameters:
      - description: Subscription payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/subscription.UpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/subscription.CUDResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
        "504":
          description: Gateway Timeout
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update subscription
      tags:
      - subscriptions
  /v1/subscriptions/totalcost:
    get:
      consumes:
      - application/json
      description: Получить суммарную стоимость подписок за период, с фильтрацией
        по пользователю и названию подписки
      parameters:
      - description: ID пользователя
        in: query
        name: user_id
        required: true
        type: string
      - description: Название подписки
        in: query
        name: service_name
        required: true
        type: string
      - description: Начало периода (MM-YYYY)
        in: query
        name: from
        required: true
        type: string
      - description: Конец периода (MM-YYYY)
        in: query
        name: to
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/subscription.TotalCostResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
        "504":
          description: Gateway Timeout
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Calculate total subscriptions cost
      tags:
      - subscriptions
swagger: "2.0"